Add another test that opens multiple tabs to force cookie issues
