+++ test-fixtures/test-cookie-store.html
@@ -43,7 +43,7 @@
   <input id="password" type="password" value="testpassword">
   <br><br>
   <button onclick="setLoginCookies()">Login</button>
-  <!-- <button onclick="clearLoginCookies()">Log Out</button> -->
+  <button onclick="clearLoginCookies()">Logout</button>
   <br><br>
   <div id="login-indicator">Checking...</div>
 </body>
